refactor: Equipmentパッケージ位置変更
refs #002
