+++ MobileSuitDatabase/src/main/java/jp/co/sunarch/mobilesuitDatabase/application/service/mobilesuit/equipment/EquipmentQueryService.java
@@ -1,13 +1,13 @@
-package jp.co.sunarch.mobilesuitDatabase.application.service.equipment;
+package jp.co.sunarch.mobilesuitDatabase.application.service.mobilesuit.equipment;
 
 import java.util.List;
 
 import org.springframework.stereotype.Component;
 
-import jp.co.sunarch.mobilesuitDatabase.application.repository.equipment.EquipmentRepository;
+import jp.co.sunarch.mobilesuitDatabase.application.repository.mobilesuit.equipment.EquipmentRepository;
 import jp.co.sunarch.mobilesuitDatabase.domain.model.arms.ArmsId;
 import jp.co.sunarch.mobilesuitDatabase.domain.model.mobilesuit.MobileSuitId;
-import jp.co.sunarch.mobilesuitDatabase.domain.model.mobilesuit.Equipment.Equipment;
+import jp.co.sunarch.mobilesuitDatabase.domain.model.mobilesuit.equipment.Equipment;
 import lombok.RequiredArgsConstructor;
 
 @Component
