+++ MobileSuitDatabase/src/main/java/jp/co/sunarch/mobilesuitDatabase/application/service/mobilesuit/equipment/EquipmentRecodeService.java
@@ -1,8 +1,8 @@
-package jp.co.sunarch.mobilesuitDatabase.application.service.equipment;
+package jp.co.sunarch.mobilesuitDatabase.application.service.mobilesuit.equipment;
 
 import org.springframework.stereotype.Component;
 
-import jp.co.sunarch.mobilesuitDatabase.application.repository.equipment.EquipmentRepository;
+import jp.co.sunarch.mobilesuitDatabase.application.repository.mobilesuit.equipment.EquipmentRepository;
 import jp.co.sunarch.mobilesuitDatabase.domain.model.arms.Arms;
 import jp.co.sunarch.mobilesuitDatabase.domain.model.mobilesuit.MobileSuit;
 import lombok.RequiredArgsConstructor;
