+++ MobileSuitDatabase/src/main/java/jp/co/sunarch/mobilesuitDatabase/application/usecase/arms/DeleteArmsUseCase.java
@@ -7,10 +7,10 @@ import org.springframework.stereotype.Service;
 import jp.co.sunarch.mobilesuitDatabase.application.command.arms.DeleteArmsCommand;
 import jp.co.sunarch.mobilesuitDatabase.application.service.arms.ArmsQueryService;
 import jp.co.sunarch.mobilesuitDatabase.application.service.arms.ArmsRecodeService;
-import jp.co.sunarch.mobilesuitDatabase.application.service.equipment.EquipmentQueryService;
-import jp.co.sunarch.mobilesuitDatabase.application.service.equipment.EquipmentRecodeService;
+import jp.co.sunarch.mobilesuitDatabase.application.service.mobilesuit.equipment.EquipmentQueryService;
+import jp.co.sunarch.mobilesuitDatabase.application.service.mobilesuit.equipment.EquipmentRecodeService;
 import jp.co.sunarch.mobilesuitDatabase.domain.model.arms.Arms;
-import jp.co.sunarch.mobilesuitDatabase.domain.model.mobilesuit.Equipment.Equipment;
+import jp.co.sunarch.mobilesuitDatabase.domain.model.mobilesuit.equipment.Equipment;
 import lombok.RequiredArgsConstructor;
 
 @Service
