+++ MobileSuitDatabase/src/main/java/jp/co/sunarch/mobilesuitDatabase/application/usecase/mobilesuit/DeleteMobileSuitUseCase.java
@@ -5,12 +5,12 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 
 import jp.co.sunarch.mobilesuitDatabase.application.command.mobilesuit.DeleteMobileSuitCommand;
-import jp.co.sunarch.mobilesuitDatabase.application.service.equipment.EquipmentQueryService;
-import jp.co.sunarch.mobilesuitDatabase.application.service.equipment.EquipmentRecodeService;
 import jp.co.sunarch.mobilesuitDatabase.application.service.mobilesuit.MobileSuitQueryService;
 import jp.co.sunarch.mobilesuitDatabase.application.service.mobilesuit.MobileSuitRecodeService;
+import jp.co.sunarch.mobilesuitDatabase.application.service.mobilesuit.equipment.EquipmentQueryService;
+import jp.co.sunarch.mobilesuitDatabase.application.service.mobilesuit.equipment.EquipmentRecodeService;
 import jp.co.sunarch.mobilesuitDatabase.domain.model.mobilesuit.MobileSuit;
-import jp.co.sunarch.mobilesuitDatabase.domain.model.mobilesuit.Equipment.Equipment;
+import jp.co.sunarch.mobilesuitDatabase.domain.model.mobilesuit.equipment.Equipment;
 import lombok.RequiredArgsConstructor;
 
 @Service
